Refactoring ExceptionsHandlingMiddleware
* the concept of dividing the middleware into several classes turned out to be non-working, since if
the exception was not caught in the first exception handler, then the other handlers could not catch
and process the exception and the client returned a response with the status code 200 OK.
