+++ serverapp/src/WebApi/Middleware/ApplicationBuilderExtensions.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 
 namespace WebApi.Middleware
 {
@@ -12,8 +6,7 @@ namespace WebApi.Middleware
     {
         public static IApplicationBuilder UseExceptionsHandlingMiddleware(this IApplicationBuilder app)
         {
-            app.UseMiddleware<ApplicationExceptionsHandlingMiddleware>();
-            app.UseMiddleware<IdentityExceptionsHandlingMiddleware>();
+            app.UseMiddleware<ExceptionsHandlingMiddleware>();
 
             return app;
         }
